+++ djiffy/urls.py
@@ -7,4 +7,6 @@ urlpatterns = [
     url(r'^(?P<id>[^/]+)/$', views.ManifestDetail.as_view(), name='manifest'),
     url(r'^(?P<manifest_id>[^/]+)/canvases/(?P<id>[^/]+)/$',
         views.CanvasDetail.as_view(), name='canvas'),
+    url(r'^canvas/autocomplete/$', views.CanvasAutocomplete.as_view(),
+        name='canvas-autocomplete'),
 ]